refactor: moved note_list to notes_container/note_list
