+++ lib/presentation/pages/home/home.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../widgets/note_list/note_list.dart';
+import '../../widgets/notes_container/note_list/note_list.dart';
 import '../note_page/note_page.dart';
 import 'bloc/bloc.dart';
