+++ lib/presentation/widgets/notes_container/note_list/bloc/note_list_bloc.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../core/error/failure.dart';
-import '../../../../core/usecases/usecase.dart';
-import '../../../../domain/usecases/get_notes.dart';
+import '../../../../../core/error/failure.dart';
+import '../../../../../core/usecases/usecase.dart';
+import '../../../../../domain/usecases/get_notes.dart';
 import 'bloc.dart';
 
 class NoteListBloc extends Bloc<NoteListEvent, NoteListState> {
