+++ lib/presentation/widgets/notes_container/note_list/bloc/note_list_state.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 
-import '../../../../domain/entities/note.dart';
+import '../../../../../domain/entities/note.dart';
 
 abstract class NoteListState extends Equatable {
   @override
