+++ lib/presentation/widgets/notes_container/note_list/note_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../injection_container.dart';
+import '../../../../injection_container.dart';
 import 'bloc/bloc.dart';
 
 class NoteList extends StatelessWidget {
